OneDrive/Desktop/tweakspeare/requirements.txt
streamlit==1.39.0
openai-whisper==20240930
pyaudio==0.2.14